Move functions related to file exclusion (reverse mode) to new file.
